fix: changing route template name issue
When I changed CaptchaImageControllerNameTemplate to a value different than "DNTCaptchaImage", then the captcha did not load (in TestWebApp). So I fixed it in ControllerRoutingConvention.

Also, I have applied some changes to prevent the action name from appearing two times in URLs: "DNTCaptchaImage/Show/Show"

The route template name (CaptchaImageControllerNameTemplate) is not required, and we can remove it, but I have not removed it because it could be used when someone wants to generate a URL of `DNTCaptchaImageController` by using `IUrlHelper.RouteUrl`.